Content Model: Support Ctrl+Backspace (1/2)
