+++ packages/roosterjs-content-model/lib/editor/corePlugins/ContentModelCopyPastePlugin.ts
@@ -1,6 +1,6 @@
 import contentModelToDom from '../../modelToDom/contentModelToDom';
 import { cloneModel } from '../../modelApi/common/cloneModel';
-import { deleteSelection } from '../../modelApi/edit/deleteSelections';
+import { deleteSelection } from '../../modelApi/edit/deleteSelection';
 import { IContentModelEditor } from '../../publicTypes/IContentModelEditor';
 import { iterateSelections } from '../../modelApi/selection/iterateSelections';
 import {
