+++ packages/roosterjs-content-model/lib/modelApi/edit/utils/DeleteSelectionStep.ts
@@ -1,0 +1,48 @@
+import { ContentModelDocument } from '../../../publicTypes/group/ContentModelDocument';
+import { ContentModelEntity } from '../../../publicTypes/entity/ContentModelEntity';
+import { ContentModelParagraph } from '../../../publicTypes/block/ContentModelParagraph';
+import { EntityOperation } from 'roosterjs-editor-types';
+import { InsertPoint } from '../../../publicTypes/selection/InsertPoint';
+import { TableSelectionContext } from '../../../publicTypes/selection/TableSelectionContext';
+import type { CompatibleEntityOperation } from 'roosterjs-editor-types/lib/compatibleTypes';
+
+/**
+ * @internal
+ */
+export interface DeleteSelectionContext {
+    insertPoint?: InsertPoint;
+    lastParagraph?: ContentModelParagraph;
+    lastTableContext?: TableSelectionContext;
+    isChanged: boolean;
+}
+
+/**
+ * @internal
+ */
+export type OnDeleteEntity = (
+    entity: ContentModelEntity,
+    operation:
+        | EntityOperation.RemoveFromStart
+        | EntityOperation.RemoveFromEnd
+        | EntityOperation.Overwrite
+        | CompatibleEntityOperation.RemoveFromStart
+        | CompatibleEntityOperation.RemoveFromEnd
+        | CompatibleEntityOperation.Overwrite
+) => boolean;
+
+/**
+ * @internal
+ */
+export interface DeleteSelectionOptions {
+    direction?: 'forward' | 'backward' | 'selectionOnly';
+    onDeleteEntity?: OnDeleteEntity;
+}
+
+/**
+ * @internal
+ */
+export type DeleteSelectionStep = (
+    context: DeleteSelectionContext,
+    options: Required<DeleteSelectionOptions>,
+    model: ContentModelDocument
+) => void;
