+++ packages/roosterjs-content-model/lib/modelApi/edit/utils/deleteSingleChar.ts
