+++ packages/roosterjs-content-model/lib/modelApi/selection/collectSelections.ts
@@ -8,12 +8,8 @@ import { ContentModelParagraph } from '../../publicTypes/block/ContentModelParag
 import { ContentModelSegment } from '../../publicTypes/segment/ContentModelSegment';
 import { ContentModelTable } from '../../publicTypes/block/ContentModelTable';
 import { isBlockGroupOfType } from '..//common/isBlockGroupOfType';
-
-import {
-    iterateSelections,
-    IterateSelectionsOption,
-    TableSelectionContext,
-} from './iterateSelections';
+import { iterateSelections, IterateSelectionsOption } from './iterateSelections';
+import { TableSelectionContext } from '../../publicTypes/selection/TableSelectionContext';
 import {
     getClosestAncestorBlockGroupIndex,
     TypeOfBlockGroup,
