+++ packages/roosterjs-content-model/test/publicApi/editing/handleBackspaceKeyTest.ts
@@ -1,4 +1,4 @@
-import * as deleteSelection from '../../../lib/modelApi/edit/deleteSelections';
+import * as deleteSelection from '../../../lib/modelApi/edit/deleteSelection';
 import * as formatWithContentModel from '../../../lib/publicApi/utils/formatWithContentModel';
 import * as handleKeyboardEventResult from '../../../lib/editor/utils/handleKeyboardEventCommon';
 import handleBackspaceKey from '../../../lib/publicApi/editing/handleBackspaceKey';
